+++ __tests__/index.js
@@ -1,5 +1,16 @@
-import * as exports from '../src';
+/* global window */
 
-test('Exports expected stuff', () => {
-  expect(exports).toMatchSnapshot();
+jest.mock('utils/webpack/require');
+
+beforeEach(() => jest.resetModules());
+
+afterEach(() => delete window.TRU_FRONT_END);
+
+test('Export at client side', () => {
+  window.TRU_FRONT_END = true;
+  expect(require('../src')).toMatchSnapshot();
+});
+
+test('Export at server side', () => {
+  expect(require('../src')).toMatchSnapshot();
 });