+++ __tests__/shared/components/Link.jsx
@@ -3,7 +3,5 @@ import React from 'react';
 import { shallowSnapshot } from 'utils/jest';
 
 test('Matches snapshots', () => {
-  shallowSnapshot((
-    <Link />
-  ));
+  shallowSnapshot(<Link />); // eslint-disable-line jsx-a11y/anchor-is-valid
 });