+++ config/test.js
@@ -1,0 +1,3 @@
+module.exports = {
+  dummyConfigKey: 'Dummy config value',
+};