+++ config/webpack/lib-base.js
@@ -35,11 +35,9 @@ module.exports = function configFactory(ops) {
     entry: ops.entry,
     externals: [
       'babel-runtime',
-      'config',
       'lodash',
       'moment',
       'prop-types',
-      'raf',
       'react',
       'react-css-super-themr',
       'react-dom',
@@ -59,6 +57,7 @@ module.exports = function configFactory(ops) {
       }),
     ],
     module: {
+      noParse: [],
       rules: [{
         /* Handles font imports in url(..) instructions in CSS. Effectively,
          * with such configuration it just rewrites those URLs to point to