+++ src/mock.js
@@ -1,4 +1,0 @@
-const mock = require('./index');
-mock.utils = require('./shared/__mocks__/utils').default;
-
-module.exports = mock;