+++ src/shared/__mocks__/utils.js
@@ -1,6 +1,0 @@
-import mock from '../utils';
-import isomorphy from '../utils/__mocks__/isomorphy';
-
-mock.isomorphy = isomorphy;
-
-export default mock;