+++ src/shared/utils/__mocks__/isomorphy.js
@@ -2,8 +2,10 @@
  * Mockup of the isomorphy module.
  */
 
+const MOCK_BUILD_TIMESTAMP = 'Wed, 29 Nov 2017 07:40:00 GMT';
+
 const mock = require.requireActual('../isomorphy');
 
-mock.buildTimestamp = () => 'Wed, 29 Nov 2017 07:40:00 GMT';
+mock.buildTimestamp = () => MOCK_BUILD_TIMESTAMP;
 
-export default mock;
+module.exports = mock;