+++ src/shared/utils/config.js
@@ -5,10 +5,12 @@
  * from node-config.
  */
 
-/* eslint-env browser */
+/* global window */
 
 import { isClientSide } from './isomorphy';
+import { requireWeak } from './webpack';
 
 /* eslint-disable global-require */
-module.exports = isClientSide() ? window.CONFIG : require('config');
+// console.log('IS CLIENT SIDE', isClientSide());
+module.exports = isClientSide() ? window.CONFIG : requireWeak('config');
 /* eslint-enable global-require */